Movement system complete

+++ src/components/rigidbody_component.hpp
@@ -4,9 +4,9 @@
 #include <glm/glm.hpp>
 
 struct RigidBodyComponent {
-	glm::vec2 velocity{};
+	glm::dvec2 velocity{};
 
-	RigidBodyComponent(glm::vec2 velocity = glm::vec2(0.0, 0.0))
+	RigidBodyComponent(glm::dvec2 velocity = glm::dvec2(0.0, 0.0))
 		: velocity{ velocity } {
 	}
 };
