+++ src/components/transform_component.hpp
@@ -4,13 +4,13 @@
 #include <glm/glm.hpp>
 
 struct TransformComponent {
-	glm::vec2 position{};
-	glm::vec2 scale{ 1.0, 1.0 };
+	glm::dvec2 position{};
+	glm::dvec2 scale{ 1.0, 1.0 };
 	double rotation{};
 
 	TransformComponent(
-		glm::vec2 position = glm::vec2(0, 0),
-		glm::vec2 scale = glm::vec2(1.0, 1.0),
+		glm::dvec2 position = glm::dvec2(0, 0),
+		glm::dvec2 scale = glm::dvec2(1.0, 1.0),
 		double rotation = 0.0)
 		: position{ position }, scale{ scale }, rotation{ rotation } {
 	}
